% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucsc_data_extraction.R
\name{closeConn}
\alias{closeConn}
\title{Close an existing connection to UCSC's public MySQL server}
\usage{
closeConn(conn)
}
\description{
Close an existing connection to UCSC's public MySQL server
}

