% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucsc_data_extraction.R
\name{fromBiomaRt}
\alias{fromBiomaRt}
\title{Extracts genes within region of interest from biomaRt.
NOTE: that biomaRt is a suggested dependecy only. So the package will run
perfectly fine if you do not have biomaRt installed. However, this
function will not execute.}
\usage{
fromBiomaRt(chromosome = NULL, start = NULL, end = NULL,
  genomeBuild = "grch37")
}
\arguments{
\item{chromosome}{int. Chromosome number [1..22]}

\item{start}{int. Start of the region, base position}

\item{end}{int. End of region, base position}

\item{genomeBuild}{string. (DEFAULT = 'grch37'). Specify the build to use. Note the build can make a dramatic difference to the returned gene set.}
}
\value{
gene set.
}
\description{
Extracts genes within region of interest from biomaRt.
NOTE: that biomaRt is a suggested dependecy only. So the package will run
perfectly fine if you do not have biomaRt installed. However, this
function will not execute.
}

