% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucsc_data_extraction.R
\name{fromUCSCEnsemblGenes}
\alias{fromUCSCEnsemblGenes}
\title{Find all genes within a defined chromosomal region.
Queries UCSC's Ensembl tables and includes additional information such as,
geneType, geneStatus and transcriptClass.
This is the best of these queries I think.}
\usage{
fromUCSCEnsemblGenes(chromosome = NULL, start = NULL, end = NULL)
}
\arguments{
\item{chromosome}{int. Chromosome number [1..22]}

\item{start}{int. Start of the region, base position}

\item{end}{int. End of region, base position}
}
\value{
query string. To be passed as argument to queryUCSC().
}
\description{
Find all genes within a defined chromosomal region.
Queries UCSC's Ensembl tables and includes additional information such as,
geneType, geneStatus and transcriptClass.
This is the best of these queries I think.
}

