% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucsc_data_extraction.R
\name{fromUCSCRefGene}
\alias{fromUCSCRefGene}
\title{Find all genes within a defined chromosomal region.
Queries UCSC's refGene table.}
\usage{
fromUCSCRefGene(chromosome = NULL, start = NULL, end = NULL)
}
\arguments{
\item{chromosome}{int. Chromosome number [1..22]}

\item{start}{int. Start of the region, base position}

\item{end}{int. End of region, base position}
}
\value{
query string. To be passed as argument to queryUCSC().
}
\description{
Find all genes within a defined chromosomal region.
Queries UCSC's refGene table.
}

