% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_ld_data.R
\name{geneAnnotation}
\alias{geneAnnotation}
\title{Adds gene annotations to an existing locus zoom, or ldZoom plot.}
\usage{
geneAnnotation(zoom, genes, geneType = "All")
}
\arguments{
\item{zoom}{GGPLOT2 plot object. See ldZoom() as example.}

\item{genes}{data frame. Genes within the target region. See fromUCSCEnsemblGenes().}

\item{geneType}{character (DEFAULT = 'All'). Filter the displayed genes by gene type (not implemented yet).}
}
\value{
zoom. A GGPLOT2 locus-zoom like plot, with gene annotations.
}
\description{
Adds gene annotations to an existing locus zoom, or ldZoom plot.
}

