% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ld_datasets.R
\name{ldByRegion}
\alias{ldByRegion}
\title{Downloads the 1000 Genomes data for the region of interest
and calculates LD between all pairs of SNPs using PLINK2.
Writes results out to ./Data/Genotype_<chromosome>_<start>_<end>.ld
Requires:
    - bash script: ./exec/getLD.sh
    - sample_list.txt must be present}
\usage{
ldByRegion(chromosome = NULL, start = NULL, end = NULL,
  download = FALSE, vcfFile = NULL, sampleFile = NULL)
}
\arguments{
\item{chromosome}{Integer Chromosome number}

\item{start}{Integer The beginning of the region of interest (in base pairs).}

\item{end}{Integer The end of the region of interest (in base pairs).}
}
\description{
Downloads the 1000 Genomes data for the region of interest
and calculates LD between all pairs of SNPs using PLINK2.
Writes results out to ./Data/Genotype_<chromosome>_<start>_<end>.ld
Requires:
    - bash script: ./exec/getLD.sh
    - sample_list.txt must be present
}

