% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_ld_data.R
\name{ldCluster}
\alias{ldCluster}
\title{Clusters SNPs based on an LD dissimilarity matrix. Returns a pvclust cluster object.
The pvclust object can be used to find:
    - similar (and dissimilar) SNPs (based on LD)
    - can be passed to pvclust::pvpick to find clusters with high/low pvalues
    - can be plotted with ldDendrogram()}
\usage{
ldCluster(ldMatrix, clustMethod = "ward.D", distMethod = "euclidean")
}
\arguments{
\item{ldMatrix}{matrix. LD dissimilarity matrix.}

\item{clustMethod}{character. Method for clustering ("average", "ward.D", "ward.D2", "single", "complete", "mcquitty", "median" or "centroid")}

\item{distMethod}{character. Method for calculating distances between SNPs. see pvclust help.}
}
\value{
pvclust cluster object.
}
\description{
Clusters SNPs based on an LD dissimilarity matrix. Returns a pvclust cluster object.
The pvclust object can be used to find:
    - similar (and dissimilar) SNPs (based on LD)
    - can be passed to pvclust::pvpick to find clusters with high/low pvalues
    - can be plotted with ldDendrogram()
}

