% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_ld_data.R
\name{ldDendrogram}
\alias{ldDendrogram}
\title{Performs heirarchical clustering of LD dissimilarity matrix and returns a dendrogram.}
\usage{
ldDendrogram(ldClusters, plotTitle = "LD-based Clustering")
}
\arguments{
\item{ldClusters}{pvclust object. Return values from ldClsuter().}

\item{plotTitle}{character (DEFAULT = LD-based Clustering). Title to be displayed with the dendrogram.}
}
\value{
dendrogram
}
\description{
Performs heirarchical clustering of LD dissimilarity matrix and returns a dendrogram.
}

