% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_ld_data.R
\name{ldDissimilarity}
\alias{ldDissimilarity}
\title{Takes a PLINK file containing pairwise R2 values between all pairs of SNPs
and converts this into an N(snp) x N(snp) dissimilarity matrix.
The dissimilarity matrix may then be used as per any other dissimilarity
matrix, e.g. clustering.}
\usage{
ldDissimilarity(ldData)
}
\arguments{
\item{ldData}{data frame. Default LD output form PLINK2. Use ldRead().}
}
\value{
ldData matrix. Dissimilarity matrixs, measured in R2.
}
\description{
Takes a PLINK file containing pairwise R2 values between all pairs of SNPs
and converts this into an N(snp) x N(snp) dissimilarity matrix.
The dissimilarity matrix may then be used as per any other dissimilarity
matrix, e.g. clustering.
}

