% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ld_datasets.R
\name{ldDownload}
\alias{ldDownload}
\title{Downloads a defined region from 1000 Genomes.}
\usage{
ldDownload(chromosome = NULL, start = NULL, end = NULL,
  outputVCF = NULL)
}
\arguments{
\item{chromosome}{integer.}

\item{start}{CHARACTER. Please note, not integer.}

\item{end}{CHARACTER. Please note, not integer.}

\item{outputVCF.}{Filename.}
}
\description{
Downloads a defined region from 1000 Genomes.
}

