% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_ld_data.R
\name{ldHeatmap}
\alias{ldHeatmap}
\title{Produces a heatmap based on an LD dissimilarity matrix.}
\usage{
ldHeatmap(ldMatrix, clustMethod = "ward.D", distMethod = "euclidean")
}
\arguments{
\item{ldMatrix}{matrix. Dissimilarity matrix of LD (R2) between SNPs.}
}
\value{
heatmap + dendrogram
}
\description{
Produces a heatmap based on an LD dissimilarity matrix.
}

