% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ld_datasets.R
\name{ldPopulation}
\alias{ldPopulation}
\title{Recursively extracts sample IDs from the 1000 Genomes panel file.
Sample IDs written to 'sample_list.txt'.}
\usage{
ldPopulation(populations, n = 1, allPops = FALSE, first = TRUE,
  popfile = system.file("extdata", "1KGPopulations.panel", package = "glida"),
  outputFile = system.file("extdata", "sample_list.txt", package = "glida"))
}
\arguments{
\item{populations}{A vector of target populations.}

\item{n}{Integer (DEFAULT = 1). Decrementor for recursion.}

\item{allPops}{Boolean (DEFAULT = FALSE). If TRUE, no filtering of samples is performed.}

\item{first}{Boolean (DEFAULT = TRUE). Controls whether to overwrite or append to existing samples file.}

\item{popfile}{Location of the 1000 Genomes panel file.}

\item{outputFile}{Location of the output file, 'sample_list.txt'.}
}
\value{
recursively iterates until all populations extracted.
}
\description{
Recursively extracts sample IDs from the 1000 Genomes panel file.
Sample IDs written to 'sample_list.txt'.
}

