% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ld_datasets.R
\name{ldProxy}
\alias{ldProxy}
\title{Given a lead SNP, calculates LD of all other SNP (relative to the lead SNP)
within a defined window (distance).
Requires:
  - bash script: ../exec/getProxy.sh
  - a VCF file (1000 Genomes data) for given region.
    Will download if not present}
\usage{
ldProxy(leadSNP, vcf = "", downloadVCF = FALSE, chromosome = NULL,
  start = NULL, end = NULL)
}
\arguments{
\item{leadSNP}{Character. The lead SNP of interest.}

\item{vcf}{filename (DEFAULT = ""). Full (or relative) path of the VCF file}

\item{downloadVCF}{Boolean (DEFAULT = FALSE). If TRUE, download the 1000 Genomes data for given region.}

\item{chromosome}{Integer (DEFAULT = NULL. If donwloadVCF == TRUE, this must be set. See ldByRegion().}

\item{start}{Integer (DEFAULT = NULL). If downloadVCF == TRUE, this must be set. See ldByRegion().}

\item{end}{Integer (DEFAULT = NULL). If downloadVCF == TRUE, this must be set. See ldByRegion().}
}
\description{
Given a lead SNP, calculates LD of all other SNP (relative to the lead SNP)
within a defined window (distance).
Requires:
  - bash script: ../exec/getProxy.sh
  - a VCF file (1000 Genomes data) for given region.
    Will download if not present
}

