% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_ld_data.R
\name{ldProxyTable}
\alias{ldProxyTable}
\title{Reads in the LD data (PLINK output) and returns an ordered summary.}
\usage{
ldProxyTable(ldProxyData)
}
\arguments{
\item{ldProxyData}{matrix. Use ldProxy() and ldRead() to generate.}
}
\value{
proxyData data frame. (Chr, Pos(A), Pos(B), SNP(A), SNP(B), R2). Ordered by R2 decreasing.
}
\description{
Reads in the LD data (PLINK output) and returns an ordered summary.
}

