% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_ld_data.R
\name{ldRead}
\alias{ldRead}
\title{Read an LD file generated by PLINK2}
\usage{
ldRead(ldFile)
}
\arguments{
\item{ldFile}{filename ("Genotype_<chromosome>_<start>_<end>.ld")}
}
\value{
ldData DataFrame (CHR_A, BP_A, SNP_A, CHR_B, BP_B, SNP_B, R2)
}
\description{
Read an LD file generated by PLINK2
}

