% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_ld_data.R
\name{ldZoom}
\alias{ldZoom}
\title{Produces a locus zoom-like plot, where the y-axis is LD (R2) rather than p-values.
LD data should be generated using ldProxy(). See relevant documentation.
Plots can be annotated with nearby genes and eQTLs (to be implemented in the future).}
\usage{
ldZoom(ldData, ldThreshold = 0.9)
}
\arguments{
\item{ldData}{data frame. Standard PLINK LD output. See ldRead().}

\item{ldThreshold}{float (DEFAULT = 0.9). SNPs names are displayed for SNPs in LD above this threshold.}
}
\value{
zoom: a ggplot2 plot object.
}
\description{
Produces a locus zoom-like plot, where the y-axis is LD (R2) rather than p-values.
LD data should be generated using ldProxy(). See relevant documentation.
Plots can be annotated with nearby genes and eQTLs (to be implemented in the future).
}

