% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucsc_data_extraction.R
\name{openConn}
\alias{openConn}
\title{Establishes a connection to UCSC's public MySQL server}
\usage{
openConn(genomeBuild = "hg19")
}
\description{
Establishes a connection to UCSC's public MySQL server
}

