% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucsc_data_extraction.R
\name{queryUCSC}
\alias{queryUCSC}
\title{Given an existing database connection, retrieve query results from UCSC}
\usage{
queryUCSC(query)
}
\arguments{
\item{query}{character. ANSI-SQL compliant query string.}
}
\value{
results data frame.
}
\description{
Given an existing database connection, retrieve query results from UCSC
}

