% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucsc_data_extraction.R
\name{updatePositions}
\alias{updatePositions}
\title{Given a vector of RSIDs, query UCSC for updated chromosomal positions}
\usage{
updatePositions(snps)
}
\arguments{
\item{snps}{vector. A vector of RSIDs.}
}
\value{
query. To be passed as argument to queryUCSC().
}
\description{
Given a vector of RSIDs, query UCSC for updated chromosomal positions
}

